chore: apply ponytail audit cleanup
