+++ package.json
@@ -94,7 +94,6 @@
     "@types/ws": "^8.18.1",
     "@xmldom/xmldom": "^0.9.11",
     "bun-types": "^1.3.10",
-    "playwright": "^1.58.2",
     "typescript": "^6.0.3"
   },
   "patchedDependencies": {
