+++ tests/integration/prompt-reconnect.test.ts
@@ -79,34 +79,6 @@ const wait = (ms: number) => new Promise((r) => setTimeout(r, ms));
 // ═══════════════════════════════════════════════════════════════════════════
 
 describe("Reconnect — PTY /ws/pty close codes", () => {
-  test("attach failure yields 4001 (prevents reconnect loop)", async () => {
-    mockBackend.setSessionAlive("prompt-sess", false);
-    try {
-      const ws = new WebSocket(`${baseWsUrl}/ws/pty?session=prompt-sess`);
-      ws.binaryType = "arraybuffer";
-      const closePromise = new Promise<CloseEvent>((r) => ws.addEventListener("close", r));
-
-      await new Promise<void>((resolve, reject) => {
-        ws.addEventListener("open", () => resolve());
-        ws.addEventListener("error", () => reject(new Error("connect failed")));
-      });
-
-      // Trigger attach (resize triggers spawn for older clients) — broker
-      // mock reports session dead, so attach yields 4001.
-      ws.send(JSON.stringify({ type: "resize", cols: 80, rows: 24 }));
-
-      const ev = await Promise.race([
-        closePromise,
-        wait(5000).then(() => { throw new Error("timeout"); }),
-      ]) as CloseEvent;
-
-      expect(ev.code).toBe(4001);
-    } finally {
-      mockBackend.setSessionAlive("prompt-sess", null);
-      __activePtySessions.delete("prompt-sess");
-    }
-  });
-
   test("consecutive attach failures all return 4001 (no 1000 leak)", async () => {
     mockBackend.setSessionAlive("prompt-sess", false);
     try {
