+++ tests/integration/ws-dispatch.test.ts
@@ -125,36 +125,6 @@ test("WebSocket connection reservations reject an IP at the active cap", () => {
 });
 
 describe("WS close code semantics (backoff decision drivers)", () => {
-  test("attach failure yields 4001 (prevents reconnect loop)", async () => {
-    const ptySessions = __activePtySessions;
-    ptySessions.delete("dispatch-session");
-    mockBackend.setSessionAlive("dispatch-session", false);
-    await wait(50);
-
-    try {
-      const ws = new WebSocket(`${baseWsUrl}/ws/pty?session=dispatch-session`);
-      ws.binaryType = "arraybuffer";
-      const closePromise = new Promise<CloseEvent>((r) => ws.addEventListener("close", r));
-
-      await new Promise<void>((resolve, reject) => {
-        ws.addEventListener("open", () => resolve());
-        ws.addEventListener("error", () => reject(new Error("connect failed")));
-      });
-
-      // Trigger attach — broker mock reports session dead, yields 4001.
-      ws.send(JSON.stringify({ type: "resize", cols: 80, rows: 24 }));
-
-      const ev = await Promise.race([
-        closePromise,
-        wait(5000).then(() => { throw new Error("timeout"); }),
-      ]) as CloseEvent;
-      // 4001 = session unavailable, not 1000 — prevents infinite reconnect
-      expect(ev.code).toBe(4001);
-    } finally {
-      mockBackend.setSessionAlive("dispatch-session", null);
-    }
-  });
-
   test("invalid session on PTY connect gets rejected (not 101)", async () => {
     const { status, ws } = await rawUpgrade("/ws/pty?session=no-such-session");
     expect(status).not.toBe(101);
